Create HUD
Created a Heads Up Display that gives the player info on health of building, total enemies in a wave, remaining enemies to destroy, and points gained. The HUD also flashed a countdown for when the wave will begin and also flashes a congratulations menu when the wave is complete.
Also upon end of the congratulations message, defeated enemies will disappear